OP-225: Refactor shop controllers
