+++ src/Resources/config/services/controller/shop.xml
@@ -2,23 +2,19 @@
 
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <services>
-        <service id="bitbag_sylius_elasticsearch_plugin.controller.action.shop.list_products" class="BitBag\SyliusElasticsearchPlugin\Controller\Action\Shop\ListProductsAction">
+        <service id="bitbag_sylius_elasticsearch_plugin.controller.action.shop.taxon_products_search" class="BitBag\SyliusElasticsearchPlugin\Controller\Action\Shop\TaxonProductsSearchAction">
             <argument type="service" id="form.factory" />
-            <argument type="service" id="bitbag_sylius_elasticsearch_plugin.controller.request_data_handler.shop_product_list" />
-            <argument type="service" id="bitbag_sylius_elasticsearch_plugin.controller.request_data_handler.shop_products_sort" />
-            <argument type="service" id="bitbag_sylius_elasticsearch_plugin.controller.request_data_handler.pagination" />
+            <argument type="service" id="bitbag_sylius_elasticsearch_plugin.controller.request_data_handler.taxon" />
             <argument type="service" id="bitbag_sylius_elasticsearch_plugin.finder.shop_products" />
             <argument type="service" id="twig" />
             <tag name="controller.service_arguments" />
         </service>
 
-        <service id="bitbag_sylius_elasticsearch_plugin.controller.action.shop.search" class="BitBag\SyliusElasticsearchPlugin\Controller\Action\Shop\SearchAction">
+        <service id="bitbag_sylius_elasticsearch_plugin.controller.action.shop.site_wide_products_search" class="BitBag\SyliusElasticsearchPlugin\Controller\Action\Shop\SiteWideProductsSearchAction">
+            <argument type="service" id="form.factory" />
+            <argument type="service" id="bitbag_sylius_elasticsearch_plugin.controller.request_data_handler.site_wide" />
+            <argument type="service" id="bitbag_sylius_elasticsearch_plugin.finder.search_products" />
             <argument type="service" id="twig" />
-            <argument type="service" id="fos_elastica.finder.bitbag_shop_product" />
-            <argument type="service" id="bitbag.sylius_elasticsearch_plugin.block_event_listener.search_form" />
-            <argument type="service" id="bitbag_sylius_elasticsearch_plugin.facet.registry" />
-            <argument type="service" id="bitbag_sylius_elasticsearch_plugin.query_builder.search_products" />
-            <argument type="service" id="bitbag_sylius_elasticsearch_plugin.controller.request_data_handler.pagination" />
             <tag name="controller.service_arguments" />
         </service>
     </services>
